+++ SU2_CFD/src/output/COutput.cpp
@@ -2161,7 +2161,7 @@ bool COutput::WriteHistoryFile_Output(const CConfig *config) {
 }
 
 bool COutput::WriteVolume_Output(CConfig *config, unsigned long Iter, bool force_writing, unsigned short iFile){
-  const unsigned short *VolumeFrequencies = config->GetVolumeOutputFrequencies();
+  const unsigned long *VolumeFrequencies = config->GetVolumeOutputFrequencies();
 
   if (config->GetTime_Domain()){
